% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variable_seurat_object_list.R
\name{add_variable_seurat_list}
\alias{add_variable_seurat_list}
\title{Takes a clinical variable vector, a character clinical variable name, and a list of seurat objects and returns the seurat object list with added metadata}
\usage{
add_variable_seurat_list(variable, variable_name, seurat_object_list)
}
\description{
Takes a clinical variable vector, a character clinical variable name, and a list of seurat objects and returns the seurat object list with added metadata
}
