% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_per_cluster.R
\name{cell_type_per_cluster}
\alias{cell_type_per_cluster}
\title{Get the count or percentage of cell type per cluster. Input is from seurat_process. output options are "count" for cell counts or "percent" as cell count percents of the cluster.}
\usage{
cell_type_per_cluster(seurat_object, output = "count")
}
\description{
Get the count or percentage of cell type per cluster. Input is from seurat_process. output options are "count" for cell counts or "percent" as cell count percents of the cluster.
}
