% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_variable_list.R
\name{clinical_variable_list}
\alias{clinical_variable_list}
\title{Creates a clinical variable list that holds clinical data  (in [\link{1}]) and the name of the variable (in [\link{2}]) as a string}
\usage{
clinical_variable_list(clinical_variable)
}
\description{
Creates a clinical variable list that holds clinical data  (in [\link{1}]) and the name of the variable (in [\link{2}]) as a string
}
