% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_of_variation.R
\name{coefficient_of_variation}
\alias{coefficient_of_variation}
\title{Computes the coefficient of variation for a group of numbers, input being similar to mean()}
\usage{
coefficient_of_variation(x)
}
\description{
Computes the coefficient of variation for a group of numbers, input being similar to mean()
}
