% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clonotype_matrix.R
\name{create_clonotype}
\alias{create_clonotype}
\title{Creates a matrix of clonotypes from 10x cellranger output}
\usage{
create_clonotype(tcr_dir, tcr_annotations)
}
\description{
Creates a matrix of clonotypes from 10x cellranger output
}
