% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_max_ggplot_build.R
\name{extract_max_ggplot_build}
\alias{extract_max_ggplot_build}
\title{Extract the max value of the y axis of a ggplot}
\usage{
extract_max_ggplot_build(p)
}
\description{
Extract the max value of the y axis of a ggplot
}
