% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cells_drops.R
\name{filter_cells_drops}
\alias{filter_cells_drops}
\title{Filter cells by barcode rank/FDR, knee, and emptydrops. Compare overlap and keep if unfiltered in at least one}
\usage{
filter_cells_drops(matrix, n_expect_cells, FDR)
}
\description{
Filter cells by barcode rank/FDR, knee, and emptydrops. Compare overlap and keep if unfiltered in at least one
}
