% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_mitochondrial_genes.R
\name{get_mitochondrial_genes}
\alias{get_mitochondrial_genes}
\title{In Seurat, mitochondrial genes are determined by the MT designation. This does not return all mitochondrial genes. We add other functions as a starting point for finding additional mitochondrial genes, but may return genes that aren't mitochondrial.}
\usage{
get_mitochondrial_genes()
}
\description{
In Seurat, mitochondrial genes are determined by the MT designation. This does not return all mitochondrial genes. We add other functions as a starting point for finding additional mitochondrial genes, but may return genes that aren't mitochondrial.
}
