% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_seurat_object_list.R
\name{integrate_seurat_object_list}
\alias{integrate_seurat_object_list}
\title{Intakes a seurat_object_list after processing via seurat_process_list and returns an integrated seurat object}
\usage{
integrate_seurat_object_list(seurat_object_list, npcs, dims)
}
\description{
Intakes a seurat_object_list after processing via seurat_process_list and returns an integrated seurat object
}
