% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dirs.R
\name{list_dirs}
\alias{list_dirs}
\title{List the directories in the path}
\usage{
list_dirs(
  path = ".",
  pattern = NULL,
  all.dirs = FALSE,
  full.names = FALSE,
  ignore.case = FALSE
)
}
\description{
List the directories in the path
}
