% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_conserved_markers_df.R
\name{make_conserved_markers_df}
\alias{make_conserved_markers_df}
\title{Find all conserved markers for each cluster and make a data frame}
\usage{
make_conserved_markers_df(seurat_object)
}
\description{
Find all conserved markers for each cluster and make a data frame
}
