% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gene_id_df.R
\name{make_gene_id_df}
\alias{make_gene_id_df}
\title{get the gene names and associated entrez ids from STARsolo output and make a dataframe}
\usage{
make_gene_id_df(star_matrix_directory, matrix_type)
}
\description{
get the gene names and associated entrez ids from STARsolo output and make a dataframe
}
