% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seurat_object.R
\name{make_seurat_object}
\alias{make_seurat_object}
\title{Makes a seurat object from a star matrix. Input is from make_star_matrix() and reference is chosen from SingleR references.}
\usage{
make_seurat_object(star_matrix, ref)
}
\description{
Makes a seurat object from a star matrix. Input is from make_star_matrix() and reference is chosen from SingleR references.
}
