% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_star_matrix_list.R
\name{make_star_matrix_list}
\alias{make_star_matrix_list}
\title{Takes a character vector from list_dirs of directories from a STAR Single Cell quantification and generates a matrix in memory}
\usage{
make_star_matrix_list(directories_vector, matrix_type)
}
\description{
Takes a character vector from list_dirs of directories from a STAR Single Cell quantification and generates a matrix in memory
}
