% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrplot_celltype.R
\name{plot_corrplot_celltype}
\alias{plot_corrplot_celltype}
\title{Plots a pie chart of percentage cell types in each cell cluster. Input is a seurat object}
\usage{
plot_corrplot_celltype(seurat_object, col)
}
\description{
Plots a pie chart of percentage cell types in each cell cluster. Input is a seurat object
}
