% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_intersect_list.R
\name{row_intersect_list}
\alias{row_intersect_list}
\title{Gets the intersection of rows between all datatables in the list}
\usage{
row_intersect_list(list)
}
\description{
Gets the intersection of rows between all datatables in the list
}
