% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_feature_plot_list.R
\name{seurat_feature_plot_list}
\alias{seurat_feature_plot_list}
\title{Intakes a seurat object list, a vector of markers, and a color col similar to ggplot and returns a list of seurat feature plots with the specified markers}
\usage{
seurat_feature_plot_list(seurat_object, markers, cols)
}
\description{
Intakes a seurat object list, a vector of markers, and a color col similar to ggplot and returns a list of seurat feature plots with the specified markers
}
