% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process_list.R
\name{seurat_process_list}
\alias{seurat_process_list}
\title{Intakes a list of seurat objects and outputs a processed list of seurat objects}
\usage{
seurat_process_list(seurat_object_list, npcs, dims)
}
\description{
Intakes a list of seurat objects and outputs a processed list of seurat objects
}
