% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process_object.R
\name{seurat_process_object}
\alias{seurat_process_object}
\title{Process a Seurat object with specified dimensions and principal components}
\usage{
seurat_process_object(seurat_object, dims, npcs)
}
\description{
Process a Seurat object with specified dimensions and principal components
}
