% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra_download.R
\name{sra_download}
\alias{sra_download}
\title{Intakes a path for an SRR_Acc_List.txt file from SRA and downloads the files into a destination directory}
\usage{
sra_download(SRR_Acc_List, destination_directory)
}
\description{
Intakes a path for an SRR_Acc_List.txt file from SRA and downloads the files into a destination directory
}
