% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_name_as_string.R
\name{variable_name_as_string}
\alias{variable_name_as_string}
\title{Returns the variable name as a string}
\usage{
variable_name_as_string(variable)
}
\description{
Returns the variable name as a string
}
