% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seurat_object_list.R
\name{make_seurat_object_list}
\alias{make_seurat_object_list}
\title{Makes a list of seurat objects. Input is from make_star_matrix() and reference is chosen from SingleR references.}
\usage{
make_seurat_object_list(matrix_list, singler_ref, singler_labels)
}
\description{
Makes a list of seurat objects. Input is from make_star_matrix() and reference is chosen from SingleR references.
}
