% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_star_matrix.R
\name{make_star_matrix}
\alias{make_star_matrix}
\title{Takes the output from a STAR Single Cell quantification and generates a matrix in memory}
\usage{
make_star_matrix(star_matrix_directory, matrix_type)
}
\description{
Takes the output from a STAR Single Cell quantification and generates a matrix in memory
}
